+++ src/main/java/org/base/advent/code2017/Day03.java
@@ -25,7 +25,7 @@ public class Day03 implements Solution<Integer> {
         return firstLargerValue(getInput());
     }
 
-    public int distanceFromOrigin(int location) {
+    public long distanceFromOrigin(int location) {
         return toPoint(location).getManhattanDistance();
     }
